tests: moved fixtures out for easy reuse
